refactor: LiveRates refactor
