+++ src/new-client/src/App/LiveRates/MainHeader/ToggleView/ChartIcon.tsx
