+++ src/new-client/src/App/LiveRates/MainHeader/ToggleView/index.ts
@@ -1,0 +1,1 @@
+export { ToggleView } from "./ToggleView"
