+++ src/new-client/src/App/LiveRates/MainHeader/index.ts
@@ -1,0 +1,1 @@
+export * from "./MainHeader"
