+++ src/new-client/src/App/LiveRates/Tiles/PriceButton/PriceButtonStyles.tsx
@@ -117,69 +117,3 @@ export const BigWrapper = styled.div`
   justify-content: center;
   flex-direction: column;
 `
-
-export const ExpiredPrice = styled.div`
-  color: ${({ theme }) => theme.colors.spectrum.uniqueCollections.Sell.base};
-  font-size: 9px;
-  text-transform: uppercase;
-`
-export const PriceControlsStyle = styled("div")<{
-  isAnalyticsView: boolean
-  isTradeExecutionInFlight?: boolean
-}>`
-  display: grid;
-  ${({ isAnalyticsView }) =>
-    isAnalyticsView
-      ? `
-      grid-template-columns: 20% 80%;
-      grid-template-rows: 50% 50%;
-      grid-template-areas: 
-      "movement sell"
-      "movement buy";
-    `
-      : `
-      grid-template-columns: 33% 33% 33%;
-      grid-template-rows: 100%;
-      grid-template-areas: 
-      "sell movement buy";
-    `}
-  ${({ isTradeExecutionInFlight }) =>
-    !isTradeExecutionInFlight && "position: relative"}
-`
-export const SellWrapper = styled.div`
-  grid-area: sell;
-`
-export const BuyWrapper = styled.div`
-  grid-area: buy;
-`
-
-export const PriceButtonDisabledPlaceholder = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-flow: column;
-  border: 2px solid ${({ theme }) => theme.core.darkBackground};
-  border-radius: 3px;
-  font-size: 10px;
-  transition: background-color 0.2s ease;
-  height: 58px;
-  min-height: 2rem;
-  max-height: 3.7rem;
-  margin-bottom: 1px;
-  min-width: 125px;
-  line-height: normal;
-  opacity: 0.5;
-  text-align: center;
-  text-transform: uppercase;
-  color: ${({ theme }) => theme.primary[5]};
-  font-weight: 400;
-`
-
-export const Icon = styled.i`
-  font-size: 20px;
-  margin: 3px 0;
-`
-
-export const AdaptiveLoaderWrapper = styled.div`
-  margin: 0 3px;
-`
