+++ src/new-client/src/App/LiveRates/Tiles/PriceButton/index.ts
@@ -1,0 +1,1 @@
+export * from "./PriceButton"
