+++ src/new-client/src/App/LiveRates/Tiles/TileHeader.tsx
@@ -1,14 +1,35 @@
-import React from "react"
-import { TileHeader as Header, TileSymbol, DeliveryDate } from "./styled"
-import { CurrencyPair } from "services/currencyPairs"
+import styled from "styled-components/macro"
+import { merge } from "rxjs"
+import { map } from "rxjs/operators"
+import { bind } from "@react-rxjs/core"
+import { format } from "date-fns"
+import { getCurrencyPair$ } from "services/currencyPairs"
+import { getPrice$ } from "services/prices"
+import { DeliveryDate } from "./styled"
 
-interface Props {
-  ccyPair: CurrencyPair
-  date: string
-}
+const [useBaseTerm, getBaseTerm$] = bind((symbol: string) =>
+  getCurrencyPair$(symbol).pipe(map(({ base, terms }) => `${base}/${terms}`)),
+)
 
-export const TileHeader: React.FC<Props> = ({ ccyPair, date }) => {
-  const baseTerm = `${ccyPair.base}/${ccyPair.terms}`
+const [useDate, getDate$] = bind((symbol: string) =>
+  getPrice$(symbol).pipe(
+    map(({ valueDate }) => `SPT (${format(new Date(valueDate), "ddMMM")})`),
+  ),
+)
+
+export const Header = styled.div`
+  display: flex;
+  align-items: center;
+`
+export const TileSymbol = styled.div`
+  color: ${({ theme }) => theme.core.textColor};
+  font-size: 0.8125rem;
+  line-height: 1rem;
+`
+
+export const TileHeader: React.FC<{ symbol: string }> = ({ symbol }) => {
+  const baseTerm = useBaseTerm(symbol)
+  const date = useDate(symbol)
   return (
     <Header>
       <TileSymbol data-qa="tile-header__tile-symbol">{baseTerm}</TileSymbol>
@@ -16,3 +37,6 @@ export const TileHeader: React.FC<Props> = ({ ccyPair, date }) => {
     </Header>
   )
 }
+
+export const tileHeader$ = (symbol: string) =>
+  merge(getBaseTerm$(symbol), getDate$(symbol))
