+++ src/new-client/src/App/LiveRates/Tiles/index.ts
@@ -1,0 +1,1 @@
+export * from "./Tiles"
