+++ src/new-client/src/App/LiveRates/index.ts
@@ -1,0 +1,1 @@
+export { LiveRates } from "./LiveRates"
