--- src/new-client/src/services/tiles.ts
@@ -1,126 +1,0 @@
-import { bind } from "@react-rxjs/core"
-import { split } from "@react-rxjs/utils"
-import { createListener } from "@react-rxjs/utils"
-import { concat, EMPTY, merge } from "rxjs"
-import {
-  distinctUntilChanged,
-  map,
-  mergeMap,
-  scan,
-  switchMap,
-  take,
-  startWith,
-  mergeAll,
-  debounceTime,
-} from "rxjs/operators"
-import { getRemoteProcedureCall$, getStream$ } from "./client"
-import { currencyPairs$, currencyPairUpdates$ } from "./currencyPairs"
-import { CamelCase, UpdateType } from "./utils"
-
-interface RawPrice {
-  Ask: number
-  Bid: number
-  Mid: number
-  CreationTimestamp: number
-  Symbol: string
-  ValueDate: string
-}
-
-export enum PriceMovementType {
-  UP,
-  DOWN,
-  NONE,
-}
-
-type HistoryPrice = CamelCase<RawPrice>
-
-export interface Price extends HistoryPrice {
-  movementType: PriceMovementType
-}
-
-interface Request {
-  symbol: string
-}
-
-export enum TileView {
-  Normal = "Normal",
-  Analytics = "Analytics",
-}
-
-export const [usePrice, getPrice$] = bind((symbol: string) =>
-  getStream$<RawPrice, Request>("pricing", "getPriceUpdates", { symbol }).pipe(
-    scan<RawPrice, Price>((acc, price) => ({
-      ask: price.Ask,
-      bid: price.Bid,
-      mid: price.Mid,
-      creationTimestamp: price.CreationTimestamp,
-      symbol: price.Symbol,
-      valueDate: price.ValueDate,
-      movementType:
-        acc === undefined
-          ? PriceMovementType.NONE
-          : price.Mid > acc.mid
-          ? PriceMovementType.UP
-          : PriceMovementType.DOWN,
-    })),
-  ),
-)
-
-const HISTORY_TIMESTAMP_THRESHOLD = 1000
-const HISTORY_SIZE = 100
-export const [useHistoricalPrices, getHistoricalPrices$] = bind<
-  [string],
-  HistoryPrice[]
->((symbol: string) =>
-  concat(
-    getRemoteProcedureCall$<HistoryPrice[], string>(
-      "priceHistory",
-      "getPriceHistory",
-      symbol,
-    ).pipe(mergeAll(), take(100)),
-    getPrice$(symbol),
-  ).pipe(
-    scan((acc, price) => {
-      if (acc.length === 0) return [price]
-      const last = acc[acc.length - 1]
-      const diff = price.creationTimestamp - last.creationTimestamp
-      if (diff < HISTORY_TIMESTAMP_THRESHOLD) return acc
-      if (acc.length < HISTORY_SIZE) return acc.concat(price)
-      const result = acc.slice(1)
-      if (price.symbol) result.push(price)
-      return result
-    }, [] as HistoryPrice[]),
-    debounceTime(0),
-  ),
-)
-
-export const tilesSubscription$ = concat(
-  currencyPairs$.pipe(
-    take(1),
-    mergeMap((ccPairs) => Object.values(ccPairs)),
-    map((currencyPair) => ({ currencyPair, updateType: UpdateType.Added })),
-  ),
-  currencyPairUpdates$,
-).pipe(
-  split(
-    (update) => update.currencyPair.symbol,
-    (update$, key) =>
-      update$.pipe(
-        distinctUntilChanged((a, b) => a.updateType === b.updateType),
-        switchMap((update) =>
-          update.updateType === UpdateType.Removed
-            ? EMPTY
-            : merge(getPrice$(key), getHistoricalPrices$(key)),
-        ),
-      ),
-  ),
-)
-
-const [selectedTileViewInput$, onSelectTileView] = createListener<TileView>()
-
-const [useSelectedTileView, selectedTileView$] = bind(
-  selectedTileViewInput$.pipe(startWith(TileView.Analytics)),
-  TileView.Analytics,
-)
-
-export { onSelectTileView, useSelectedTileView, selectedTileView$ }
